More refactoring of config generate subcommand
